Move Thrift metastore classes to separate package
