+++ presto-hive/src/main/java/com/facebook/presto/hive/metastore/thrift/HiveMetastoreApiStats.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.hive.metastore;
+package com.facebook.presto.hive.metastore.thrift;
 
 import io.airlift.stats.CounterStat;
 import io.airlift.stats.TimeStat;
