+++ presto-hive/src/main/java/com/facebook/presto/hive/metastore/thrift/HiveMetastoreClientFactory.java
@@ -11,11 +11,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.hive;
+package com.facebook.presto.hive.metastore.thrift;
 
+import com.facebook.presto.hive.HiveClientConfig;
 import com.facebook.presto.hive.authentication.HiveMetastoreAuthentication;
-import com.facebook.presto.hive.metastore.HiveMetastoreClient;
-import com.facebook.presto.hive.thrift.Transport;
 import com.google.common.net.HostAndPort;
 import io.airlift.units.Duration;
 import org.apache.thrift.transport.TTransportException;
