+++ presto-hive/src/test/java/com/facebook/presto/hive/metastore/thrift/MockHiveMetastoreClientFactory.java
@@ -11,10 +11,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.hive;
+package com.facebook.presto.hive.metastore.thrift;
 
 import com.facebook.presto.hive.authentication.NoHiveMetastoreAuthentication;
-import com.facebook.presto.hive.metastore.HiveMetastoreClient;
 import com.google.common.net.HostAndPort;
 import io.airlift.units.Duration;
 import org.apache.thrift.transport.TTransportException;
