+++ presto-hive/src/test/java/com/facebook/presto/hive/metastore/thrift/TestingHiveCluster.java
@@ -11,10 +11,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.hive;
+package com.facebook.presto.hive.metastore.thrift;
 
+import com.facebook.presto.hive.HiveClientConfig;
 import com.facebook.presto.hive.authentication.NoHiveMetastoreAuthentication;
-import com.facebook.presto.hive.metastore.HiveMetastoreClient;
 import com.google.common.base.Throwables;
 import org.apache.thrift.transport.TTransportException;
